// apps/web/src/lib/api/generated/main/model/heroCoverageGapResponseDto.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { HeroCoverageGapResponseDtoGapType } from './heroCoverageGapResponseDtoGapType';
import type { HeroCoverageGapResponseDtoMap } from './heroCoverageGapResponseDtoMap';

export interface HeroCoverageGapResponseDto {
  id: string;
  mapId: string;
  heroNpcId: string;
  gapType: HeroCoverageGapResponseDtoGapType;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  startedAt: string;
  /**
     * @nullable
     * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
     */
  endedAt: string | null;
  /** @nullable */
  durationSeconds: number | null;
  map: HeroCoverageGapResponseDtoMap;
}